requirements.txt
cryptography==1.5.2
requests==2.11.1
pyOpenSSL==16.2.0
